reorganize materials
