+++ materials/foggy/fragment.glsl
