+++ style-3d.js
@@ -86,38 +86,5 @@ module.exports = {
   },
 
   '.inactive': {
-    emissive: [0.1, 0.1, 0.1]
   }
 }
-
-// module.exports = {
-//   player: {
-//     color: [0.9, 0.9, 0.9],
-//     decay: 100,
-//     bright: 0
-//   },
-
-//   cue0: {
-//     color: [0.87, 0.52, 0.23],
-//     decay: 30,
-//     bright: 8
-//   },
-
-//   cue1: {
-//     color: [0.00, 0.76, 0.93],
-//     decay: 30,
-//     bright: 8
-//   },
-
-//   cue2: {
-//     color: [0.51, 0.79, 0.29],
-//     decay: 30,
-//     bright: 8
-//   },
-
-//   cue3: {
-//     color: [0.81, 0.33, 0.34],
-//     decay: 30,
-//     bright: 8
-//   },
-// }
